# doc/swagger.yaml
swagger: '2.0'
info:
  description: Simple demo crud server with cats
  version: 1.0.0
  title: Swagger Cats
  contact:
    email: jurij.kaminskij@gmail.com
host: '127.0.0.1:8080'
basePath: /
schemes:
  - http
paths:
  /cats:
    get:
      tags:
        - cats
      summary: List cats with parameters filtration
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/gender'
        - $ref: '#/parameters/color'
        - $ref: '#/parameters/vaccinated'
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: '#/definitions/Cat'
  /cats/add:
    post:
      tags:
        - cats
      summary: Insert new cat
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: Cat object
          required: true
          schema:
            $ref: '#/definitions/Cat'
      responses:
        '200':
          description: successful operation
  '/cats/{catName}':
    get:
      tags:
        - cats
      summary: Get cat by uniq name
      produces:
        - application/json
      parameters:
        - name: catName
          in: path
          description: Uniq name of the cat
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Cat'
    delete:
      tags:
        - cats
      summary: Delete cat by uniq name
      produces:
        - application/json
      parameters:
        - name: catName
          in: path
          description: Uniq name of the cat
          required: true
          type: string
      responses:
        '200':
          description: successful operation
    put:
      tags:
        - cats
      summary: Update cat by uniq name and url parameters
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/catName'
        - $ref: '#/parameters/gender'
        - $ref: '#/parameters/color'
        - $ref: '#/parameters/vaccinated'
      responses:
        '200':
          description: successful operation
definitions:
  Cat:
    type: object
    required:
      - name
      - gender
      - color
      - vaccinated
    properties:
      name:
        type: string
        example: Felix
      gender:
        type: string
        enum:
          - male
          - female
      color:
        type: string
        example: black
      vaccinated:
        type: boolean
parameters:
  catName:
    name: catName
    in: path
    description: Uniq name of the cat
    required: true
    type: string
  gender:
    name: gender
    in: query
    type: string
    enum:
      - male
      - female
  color:
    name: color
    in: query
    type: string
  vaccinated:
    name: vaccinated
    in: query
    type: boolean